requirements.txt
Django==5.0.1
djangorestframework==3.15.2
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.28.0
gunicorn==23.0.0
psycopg2-binary
whitenoise==6.6.0  